requirements.txt
# generated from manifests external_dependencies
odoo_test_helper
openupgradelib
